[PYTHON] Factorizing exec call + DataStrategy class.
code facotrization done with call to subprocess.check_output in
Profiling and Strategy to execute in Envvars.

Adding renaming of plt and chk files for PMF simulation.
TODO: This should be done as an option.

Implementation of execAllStrat option: the execution of different
strategies does not stop when a valid one is found.

Developping DataStrategy class to contain the different strategy we
would like to test: For now they are specific to PeleC/PMF with  30
timesteps simulation.