[C++] Fix missing index bug
Some indexes were missing in trace (both profile and strategy).
This was due to mapStatic (dladdr/backtrace_symbols based) and mapDynamic (ASLR sensitive/unwind/backtrace based)
not being always equal, as this was assume in the code (without assert ....)
Once in a while a dyncall would not exist in mapDynamic but exist in mapStatic
thus erasing the element already present in the latter (because it was assume this callstack was encountered for .
the first time.