[PYTHON] First bricks for MultiStepSiteSearch.
Nothing has been tested yet.
Need to find the link between the bricks, it may
be in success and failure set: maybe a list of size
number of backtraces, keep the type configuration for all of them,
and update it at each stage... Like in the paper.
Only missing thing in the paper is the application profile, which is not mentioned.